Refactor component imports and styles; update CSS file structure for better organization.
