+++ src/styles/Footer.jsx
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 
-export const FooterStyles = styled.footer`
+export const Footer = styled.footer`
   display: flex;
   flex-direction: row;
   justify-content: center;
